AppRTCDemo: replace the use of query-string parameters for pre-JB devices.
Replaces the use of a query-string parameter with a (once-per-session)
JS-to-Java function call, because query-string parameters on file:// URLs are
busted on ICS and earlier Android releases
(https://code.google.com/p/android/issues/detail?id=17535).

Also added channel.html to the list of inputs to cause edits to it to cause a
rebuild of the .apk.

BUG=1949
R=wu@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/1890004

git-svn-id: http://webrtc.googlecode.com/svn/trunk/talk@4421 4adac7df-926f-26a2-2b94-8c16560cd09d