PeerConnectionFactory: delay deletion of owned threads.
Since PeerConnection holds a ref to its creating PeerConnectionFactory, it's
possible for ~PeerConnectionFactory() to be run on its signaling thread.
Deleting a thread from within that thread is sad times, so don't do it.

It would be nicer to avoid having PeerConnection hold a ref to the factory,
and instead require the user to keep the factory alive.  Unfortunately that
changes the contract on PeerConnection{,Factory} and it's unclear how to vet
existing callers for safety.

BUG=3100
R=juberti@webrtc.org, noahric@google.com

Review URL: https://webrtc-codereview.appspot.com/11289004

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: cea024d6728c5bc897b542b90de5f55e75cf3fbd