Refactor StatsCollector and associated types. * Due to the type changes, I'm going to update the OnCompleted event in two phases to sync with Chrome. This is the first phase. * Reports are now managed in a set, not a map, since it's enough to store the id in one place. * Report ids are now const. * Copying of data has been greatly reduced. * This change includes preparation work for making GetStats fully async.
This is a reland of r6778 which was reverted due to fyi bots failing.
I found and fixed the issue which was that in a couple of places I needed to replace a report instead of finding+updating an existing one.

R=xians@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/15119004

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 730bf30da75514f22fc1869a93e130e582a8e045