Remove flake in SendsLowerResolutionOnSmallerFrames.
Speculative fix for break on Linux64 Release. It looks like the second
frame is being dropped which is likely because the two frames are sent
too close to eachother. Adding a delay of 33ms in between them to make
sure the second one isn't dropped.

R=minyue@webrtc.org
TBR=pthatcher@webrtc.org
BUG=1788

Review URL: https://webrtc-codereview.appspot.com/22289004

git-svn-id: http://webrtc.googlecode.com/svn/trunk/talk@7043 4adac7df-926f-26a2-2b94-8c16560cd09d