AppRTCDemoActivity: use differnet Themes for different API levels
The current AndroidManifest.xml hardcodes a Theme that
is only available in Android L or later (Material). To
maintain backwards compatibility, and for better App
style, create a single Theme/Style and define it for
different APIs.

I tested this in two Nexus %, one with prerelease L
and another with a KK 4.4.2 and the Theme is indeed
automagically updated :)

Note that this is compatible with
https://webrtc-codereview.appspot.com/26979004/

R=glaznev@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/26019004

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 8944c9d08b6585b25161ecafc60014e158ef3a25