Use a variant for storing stats values in StatsCollector code.
This cuts down on the amount of string copying we currently do and paves the way for separating the code that fetches the stats from the code that populates the stats reports.  As is, that code is intertwined, so we populate the stats on both signaling and worker thread.

I'm also adding some documentation and TODOs for further improvements.

BUG=2822
R=pthatcher@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/47459004

Cr-Original-Commit-Position: refs/heads/master@{#8700}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: d3900296ae4416de2ea21be4548ea4adba8f3280