Disables SW AEC when built-in AEC is enabled
As of r7849 the built-in AEC on devicing supporting it is enabled by default.
Unfortunately, the SW AEC (AECM) was not disabled, hence running on top of the built-in one. This is not necessary. In fact it reduce double talk performance significantly.

BUG=4431
TESTED=manually
R=henrika@webrtc.org, mflodman@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/49419004

Cr-Original-Commit-Position: refs/heads/master@{#8735}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 3f11823a1a802d6073c416d32c347e7fb6b236f7