JavaVideoRendererWrapper: Use jlongFromPointer() to convert frame pointer to jlong
The purpose of this CL is to use jlongFromPointer() for converting frame pointers to jlong instead of implicit casts which is not safe.

In order to respect constness, I had to make a small helper function for this.

BUG=webrtc:4993
R=perkj@webrtc.org

Review URL: https://codereview.webrtc.org/1373233002 .

Cr-Original-Commit-Position: refs/heads/master@{#10108}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 1d640e53bde26f6951e274a0837aeedb82f5ea3f